Fix problem with USB mouse/keyboard driver
Under TOS, the USB mouse and keyboard drivers are called from the
timer interrupt via etv_timer, and therefore run with interrupts
disabled.  If the ikbd is sending data to the system at this time,
keystrokes or parts of mouse packets may be dropped, leading to
ugly system behaviour.
Thanks to Ingo Uhlemann/Christian Zietz for their code in the
Lightning driver which is the basis of this fix.